Extract channelState fixture to its own file
