+++ packages/server-wallet/src/handlers/__test__/updateChannel.test.ts
@@ -1,5 +1,6 @@
 import {updateChannel} from '../updateChannel';
-import {updateChannelFixture, channelStateFixture} from '../fixtures/updateChannel';
+import {updateChannelFixture} from '../fixtures/updateChannel';
+import {channelStateFixture} from '../../protocols/__test__/fixtures/channel-state';
 
 test('validUpdate', () => {
   updateChannel(updateChannelFixture(), channelStateFixture());
