+++ packages/server-wallet/src/protocols/__test__/fixtures/channel-state.ts
@@ -1,0 +1,14 @@
+import {ChannelState} from '../../state';
+import {stateWithHashSignedBy} from '../../../wallet/__test__/fixtures/states';
+import {Uint256} from '../../../type-aliases';
+import {fixture} from '../../../wallet/__test__/fixtures/utils';
+
+const defaultChannelState: ChannelState = {
+  channelId: '0x1234',
+  myIndex: 0,
+  supported: stateWithHashSignedBy()({turnNum: 3}),
+  latest: stateWithHashSignedBy()({turnNum: 3}),
+  funding: (): Uint256 => '0x0',
+};
+
+export const channelStateFixture = fixture(defaultChannelState);
